refactor(query): share aggregation execution paths
