+++ wow-elasticsearch/src/main/kotlin/me/ahoo/wow/elasticsearch/query/snapshot/ElasticsearchSnapshotQueryService.kt
@@ -144,16 +144,7 @@ class ElasticsearchSnapshotQueryService<S : Any> private constructor(
     override fun resolve(filter: FilterExpression) = schemaProvider.resolve(filter, validationMode)
 
     override fun aggregate(query: AggregationQuery): Flux<DynamicDocument> =
-        schemaProvider.resolve(query, validationMode).flatMapMany { resolved ->
-            ElasticsearchAggregationPager(
-                elasticsearchClient,
-                indexName,
-                queryBatchSize,
-                queryKeepAlive,
-            ).execute(
-                ElasticsearchAggregationCompiler(filterConverter).compile(resolved.query, resolved.schema),
-            )
-        }
+        schemaProvider.resolve(query, validationMode).flatMapMany(::executeAggregation)
 
     companion object {
         private fun defaultSchemaProvider(
